pyproject.toml
[project]
name = "pip-zendure"
version = "0.1.0"
description = "API client for Zendure devices"
authors = [{ name = "Daniel Tschinder" }]
readme = "README.md"
requires-python = ">=3.13.0,<4.0"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "aiomqtt (>=2.3.0,<3.0.0)",
    "httpx-retry (>=0.2.1,<0.3.0)",
    "pycryptodome (>=3.21.0,<4.0.0)",
    "usingversion (>=0.1.2,<0.2.0)",
]

[tool.poetry]
packages = [{ include = "pip_zendure", from = "src" }]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
